feat(ui): Improve Stage Details view
