+++ ui/src/app.less
@@ -83,3 +83,7 @@ body, html {
 .ant-tabs-content {
     height: 100%;
 }
+
+.ant-drawer-close {
+    margin-inline-end: 12px !important;
+}
