--- ui/src/features/stage/requested-freight.module.less
@@ -1,8 +1,0 @@
-.active-freight-item {
-  border: 2px solid #66b2ff;
-  box-shadow: 0 4px 12px rgba(0, 123, 255, 0.2), 0 2px 4px rgba(0, 0, 0, 0.1);
-
-  &:hover {
-    border: 2px solid #66b2ff !important;
-  }
-}
