+++ ui/src/features/stage/tabs/settings/stage-edit-form.tsx
@@ -1,13 +1,13 @@
 import { useMutation, useQuery } from '@connectrpc/connect-query';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { Button, Modal, Space, Typography } from 'antd';
+import { Button, Flex, message, Space, Typography } from 'antd';
 import type { JSONSchema4 } from 'json-schema';
 import { useForm } from 'react-hook-form';
+import { useParams } from 'react-router-dom';
 import { z } from 'zod';
 
 import { YamlEditor } from '@ui/features/common/code-editor/yaml-editor';
 import { FieldContainer } from '@ui/features/common/form/field-container';
-import { ModalComponentProps } from '@ui/features/common/modal/modal-context';
 import {
   getStage,
   updateResource
@@ -17,18 +17,14 @@ import schema from '@ui/gen/schema/stages.kargo.akuity.io_v1alpha1.json';
 import { decodeRawData } from '@ui/utils/decode-raw-data';
 import { zodValidators } from '@ui/utils/validators';
 
-import { getStageYAMLExample } from '../project/pipelines/utils/stage-yaml-example';
-
-type Props = ModalComponentProps & {
-  projectName: string;
-  stageName: string;
-};
+import { getStageYAMLExample } from '../../../project/pipelines/utils/stage-yaml-example';
 
 const formSchema = z.object({
   value: zodValidators.requiredString
 });
 
-export const EditStageModal = ({ visible, hide, projectName, stageName }: Props) => {
+export const StageEditForm = () => {
+  const { name: projectName, stageName } = useParams();
   const { data, isLoading } = useQuery(getStage, {
     project: projectName,
     name: stageName,
@@ -36,7 +32,7 @@ export const EditStageModal = ({ visible, hide, projectName, stageName }: Props)
   });
 
   const { mutateAsync, isPending } = useMutation(updateResource, {
-    onSuccess: () => hide()
+    onSuccess: () => message.success('Stage has been updated')
   });
 
   const { control, handleSubmit } = useForm({
@@ -54,26 +50,7 @@ export const EditStageModal = ({ visible, hide, projectName, stageName }: Props)
   });
 
   return (
-    <Modal
-      destroyOnClose
-      open={visible}
-      title='Edit Stage'
-      closable={false}
-      width={680}
-      footer={
-        <div className='flex items-center justify-between'>
-          <Typography.Link href='https://docs.kargo.io/quickstart/#the-test-stage' target='_blank'>
-            Documentation
-          </Typography.Link>
-          <Space>
-            <Button onClick={hide}>Cancel</Button>
-            <Button type='primary' onClick={onSubmit} loading={isPending}>
-              Update
-            </Button>
-          </Space>
-        </div>
-      }
-    >
+    <>
       <FieldContainer name='value' control={control}>
         {({ field: { value, onChange } }) => (
           <YamlEditor
@@ -81,7 +58,7 @@ export const EditStageModal = ({ visible, hide, projectName, stageName }: Props)
             onChange={(e) => onChange(e || '')}
             height='500px'
             schema={schema as JSONSchema4}
-            placeholder={getStageYAMLExample(projectName)}
+            placeholder={projectName && getStageYAMLExample(projectName)}
             isLoading={isLoading}
             isHideManagedFieldsDisplayed
             label='YAML'
@@ -89,6 +66,17 @@ export const EditStageModal = ({ visible, hide, projectName, stageName }: Props)
           />
         )}
       </FieldContainer>
-    </Modal>
+
+      <Flex justify='space-between'>
+        <Typography.Link href='https://docs.kargo.io/quickstart/#the-test-stage' target='_blank'>
+          Documentation
+        </Typography.Link>
+        <Space>
+          <Button type='primary' onClick={onSubmit} loading={isPending}>
+            Update
+          </Button>
+        </Space>
+      </Flex>
+    </>
   );
 };
